pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-document-handler"
version = "0.1.0"
description = "Document management system with smart URL organization and RAG capabilities for Milvus database"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Tony Philip"},
]
keywords = ["rag", "milvus", "document-management", "url-management", "web-scraping", "langchain", "flask"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

license = "MIT"

dependencies = [
    "flask>=2.3.0",
    "pymilvus>=2.3.0",
    "sentence-transformers>=2.2.0",
    "werkzeug>=2.3.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
    "python-docx>=0.8.11",
    "chardet>=5.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-unstructured>=0.1.0",
    "unstructured[pdf]>=0.10.0",
    "cryptography>=41.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "exchangelib>=5.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/tonyphilip/rag-document-handler"
Repository = "https://github.com/tonyphilip/rag-document-handler"
Documentation = "https://github.com/tonyphilip/rag-document-handler#readme"

[tool.setuptools]
packages = ["rag_manager", "rag_manager.core", "rag_manager.processors", "rag_manager.managers", "rag_manager.web", "ingestion"]

[project.scripts]
rag-document-handler = "main:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
